Remove docker from QGIS tests (#943)
The tests are now run via the qgis_testrunner script that is a clone
from the original QGIS runner:
https://github.com/qgis/QGIS/blob/4891c6bda8922ff5c6a70021f097ce43c486cddf/.docker/qgis_resources/test_runner/qgis_testrunner.py.
I removed the if check for being inside or outside of QGIS, because not
every module was available in Windows.

The QGIS profile information and the plugins themselves are now
installed inside of the .pixi folder, so that cleanup is easier. And
there is no need anymore for looking into the roaming folder.

The QGIS tests that really require the user interface are called via
`test-ribasim-qgis-ui`. This task is not automated.
The QGIS tests that do not require the user interface are called via
`test-ribasim-qgis` and are automated via github actions.

---------

Co-authored-by: Maarten Pronk <git@evetion.nl>
Co-authored-by: Maarten Pronk <8655030+evetion@users.noreply.github.com>