Support only fixed saving intervals, set default at daily (#1155)
The default `saveat` was to save every timestep. This is nice for small
test models to see exactly what is happening, but not great for any
serious model, where it quickly leads to very large output files. Users
that didn't know about `saveat` had trouble processing results due to
the large size. Daily may be too coarse for some applications, but it is
less likely to lead to such issues.

This also removes a feature that nobody asked for or, to my knowledge,
used. You could set `saveat` with a list to save results at irregular
intervals:

```toml
[solver]
saveat = [3600, 3660]
```

The default of the empty list was sometimes confusing, and the schema
was a bit odd supporting either a float or list of floats. We copied
SciML here 1:1, but I think it is clearer to let this go, and require
either fixed intervals or every timestep with `saveat = 0`.