add realized user demand and expose via BMI (#1190)
Fixes https://github.com/Deltares/Ribasim/issues/1098.

@visr @HendrikKok would it be OK if `realized` is not an average but a
cumulative volume and resetting it to `0.0` is the responsibility of the
coupler? Otherwise we need a discussion about where in the Ribasim code
these things must happen in such a way that dividing by the time
interval happens before `realized` is accessed and resetting afterwards.

Also good to know that `demand` is per user per priority, but `realized`
is only per user.

**For BMI variable name upgrade instructions see:**


https://github.com/Deltares/Ribasim/pull/1190/files#diff-88d699c34d26c16edb268bdacbda2afb8dad704b8db2931a83e04b9ca52dc31a.

---------

Co-authored-by: Martijn Visser <mgvisser@gmail.com>