Fix precompile script (#1211)
Taken out of #1204 whilst that is on hold.

The directories changed in #1021, causing our precompile run to return
error code 1 with a "TOML not found error". That means we were not
precompiling an actual run and therefore probably need to wait a bit
longer before the start of a simulation. This fixes the directory and
adds an assertion to make sure the simulation is successful.