QGIS plugin updates for `node_type`, remove `explode_and_connect` (#1220)
@jvanhouwelingen noticed that we don't create a `node_id` column instead
of `fid`. To be able to edit tables in QGIS we always need `fid`, but
its value no longer matters, so leaving it to autogenerate is normally
best.

This removes `explode_and_connect` to automatically fill in part of the
Edge attributes when creating a new geometry. It was no longer correct
due to the schema updates, so better to remove it for now, and accept
that users have to both draw the geometry and fill in the correct
attributes. An issue with the desired behavior is already written in
#352. In short, `derive_connectivity` would still be nice to have, but
not `explode`.

- add `node_id` to the Node schema
- use `node_id` in the Node label
- add `from_node_type` and `to_node_type` to the Edge schema
- keep showing fid in the Edge label to be able to associate edge_id in
timeseries widget
- use a dropdown selector for the Node and Edge editor widget, see
screenshot below
- update schemas with `listen_node_type` as in #1110


![image](https://github.com/Deltares/Ribasim/assets/4471859/aae55ec1-0612-408f-864b-e1158af8fe28)

Edit: with the updates from Huite this fixes #352

---------

Co-authored-by: Huite Bootsma <huitebootsma@gmail.com>