Add plotting support for multiple line segment edges (#1223)
Fixes #1218

test_edge_plot:

![image](https://github.com/Deltares/Ribasim/assets/13662783/a6d270c5-407e-4234-9e27-001499273500)

And one of the testmodels I edited in QGIS to add additional vertices:

![image](https://github.com/Deltares/Ribasim/assets/13662783/58ff9234-6aa9-46b1-b9d6-b1c9b058ef77)

I also took the opportunity to use `.pop()` on the kwargs dicts.

The marker colors also had to be adjusted since there's no longer a 1:1
marked:edge_geometry relationship. I'm using a numpy array instead of
generating the list, since the arrays support indexing.
Using boolean indexing for setting the colors doesn't work very well
because the lengths of the color names might differ, and need to be
supplied as the dtype. `np.where` is clever enough to take care of it,
generating the right dtype on the fly.