Use 32 bit signed integer for all IDs and priorities (#1252)
We used to have a test that `typemax(Int64)` was working as a node ID,
but this broke in the add API. Rather than fixing that I think Int32 IDs
will always suffice (goes to over 2 billion), and save some memory and
(uncompressed) output file size.

Also when creating a new table in QGIS it creates Int32:

![image](https://github.com/Deltares/Ribasim/assets/4471859/c12f2baf-bc84-4086-8696-27c7e6d0d9fb)

Diff is fairly large since we are restrictive in our function type
signatures. `Int` is still used for indices.