Don't save states on callbacks (#1281)
Closes #1213, in the sense that I looked into the issue and it does not
occur on the last release.
I did notice that `basin.arrow` had multiple entries for the same
timestamp on a transient run with daily forcing. The
`PresetTimeCallback` updating the forcing caused these extra saves. The
default `save_positions=(true,true)` is needed when modifying the state,
but we don't do that. I checked all callbacks for this behavior.

This does mean that a few tests had to be modified because we don't add
an extra save if we hit a DiscreteControl condition, but have to wait
until the next state save.


https://docs.sciml.ai/DiffEqDocs/stable/features/callback_functions/#SciMLBase.DiscreteCallback