Set the Node fid back to the node_id (#1316)
The idea was that the Node table `fid` would not be used anymore, but
#1306 showed that is unfortunately not yet the case. So this changes the
`fid` after it is initially written to be equal to the `node_id`. Any
users should still not rely on the `fid`, because this is only
temporary.

This is like #1311 but now also catching non unique node IDs on
`model.write`.
It also provides a temporary fix to #1306. It's not great, but I feel
that especially #1306 is bad enough to warrant a quick fix. This
would've been a bit cleaner to implement using #1312 to avoid having to
read and re-write the Node table, but that PR still has issues we don't
understand, so best not to wait on that.


![image](https://github.com/Deltares/Ribasim/assets/4471859/3efc2f8c-6be5-459d-929a-7ce8568fca3d)