No need to formulate flow for Terminal anymore (#1357)
This was missed from #1300. We don't store vertical flows anymore. Doing
this was the only reason for `formulate_flows!` for Terminal nodes.

What I don't fully understand is why tests did not fail on this. Any
test model with a Terminal should have failed since the 3-arg
`add_flow!(graph, id, -q)` it called was removed in #1300. This PR also
removes the 4-arg version that is no longer used.