Don't add the endtime to basin.arrow results (#1359)
This reverts a part of 4ff9987710fac6642d723d97bceb574517b09f24

On Tuesday we discussed the `basin.arrow` with @SouthEndMusic and
@Huite. A few issues/PRs will come out of this, but this already removes
the endtime, such that we don't have `missing` vertical fluxes in
`basin.arrow` anymore. The downside is that the final storage is not
currently available, but we have #1358 for that, and plan to add `dS/dt
(m3/s)` as well to be able to compute it.