Add remaining water balance terms to basin.arrow (#1367)
Technically these terms can be calculated from flow.arrow and the
storage. But since calculating the water balance for a Basin is such a
common operation, we include them directly. This also makes it easier to
locate water balance errors in time and space, by including the error
term.

At some point we may also want to add the instantaneous inflow and
outflow to the Basin struct, and update them in `formulate_du!`. This
would allow us to reject steps with too large of a balance error, using
one of the metrics mentioned in #1271. But for now this only adds them
to the output via a callback.

---------

Co-authored-by: Bart de Koning <74617371+SouthEndMusic@users.noreply.github.com>