Sort only Edge fid so edge_id becomes more stable (#1363)
Before we sorted the Edge table like this:

```python
        sort_keys = [
            "from_node_type",
            "from_node_id",
            "to_node_type",
            "to_node_id",
        ]
```

This made it appear a bit more neat, though it served no other purpose.

The `fid` index was mostly an implementation detail that the user did
not specify, and it went from 1:n following the sorting above. However
`fid` becomes `edge_id` in the flow.arrow. Therefore when a users added
a new edge, usually half of all `edge_id`s changed, making
post-processing unnecessarily difficult. Therefore this PR removes this
sorting, such that the input order is retained, keeping the old
`edge_id`s stable.

With this I think we can close #1310. It is useful to have a single
identifier value for an Edge, even though it is somewhat superfluous.
With this PR it becomes stable, unless users start modifying `fid`s
themselves.