Switch to TimeZones release 1.15 (#1432)
Since [our
contribution](https://github.com/JuliaTime/TimeZones.jl/pull/450) is
merged upstream and released, we can stop using a fork.
We don't directly depend on this package, but it is an Arrow.jl
dependency. Therefore it can now be removed as a direct dependency,
which was only done to be able to specify compat and point it to a fork.