Move to DataInterpolations v5
This was breaking for us since it removed a type alias that we defined in our const alias ScalarInterpolation and VectorInterpolation.

This also tightens some return types to those concrete aliases, and changes a LevelDemand field type from the abstract `Vector{LinearInterpolation}` to the concrete `Vector{ScalarInterpolation}`, which should help the compiler.