Offline coupling with Delwaq (#1137)
Fixes #649

This PR adds a folder `coupling/delwaq` for offline coupling with
Delwaq, with a more extensive description provided in `README.md` in the
same folder.

A test is provided in `test.py`, that is also run on Teamcity on
Windows, using the weekly DIMR releases. The `basic` testmodel (the only
with built-in substance concentration values) is used for this. In order
to speed-up the CI one can now do `pixi run ribasim-core-testmodels
basic` to only run the `basic` model (but any or multiple model names
can be used).

---------

Co-authored-by: Martijn Visser <mgvisser@gmail.com>
Co-authored-by: Jingru Feng <feng_jingru@foxmail.com>