Convert TeamCity Xml configuration files to Kotlin (#1546)
Fixes #

This commit replaces the the TeamCity configuration files written in xml
with the newer format written in Kotlin.
This has several advantages:
- Make the configuration files easier to read
- Makes it easier to recognize duplicate code and extract them to
templates
- Makes it possible to share template between teams that also uses
Kotlin
- Makes it possible to write more complex configurations because the
configuration files can now contain logic

The current versioned settings of Ribasim are set to the xml format.
Therefor I couldn't test these changes in the existing Ribasim project.
Instead I created a separate Ribasim_Sandbox project that is set to use
the Kotlin format. The build seems ok. The reason I say ok is because
I'm not able to test everything due to missing credentials which you
cant copy between projects

When merging this commit to main we have to set the settings format in
the versioned settings tab to Kotlin otherwise the builds will fail