Use keywords for creating parameter structs (#1566)
By putting the
[kwdef](https://docs.julialang.org/en/v1/base/base/#Base.@kwdef) macro
in front of our struct definitions we can create them with keywords and
set defaults if applicable. This helps with readability and makes it
less likely to pass the wrong argument due to a wrong order. E.g.
compare this test code before and after:

```julia
basin = Ribasim.Basin(
    NodeID.(:Basin, [1], 1),
    [NodeID[]],
    [NodeID[]],
    zeros(1),
    zeros(1),
    zeros(1),
    zeros(1),
    zeros(1),
    zeros(1),
    zeros(1),
    [storage_to_level],
    [level_to_area],
    demand,
    StructVector{Ribasim.BasinTimeV1}(undef, 0),
)
```

```julia
basin = Ribasim.Basin(;
    node_id = NodeID.(:Basin, [1], 1),
    storage_to_level = [storage_to_level],
    level_to_area = [level_to_area],
    demand,
    time = StructVector{Ribasim.BasinTimeV1}(undef, 0),
)```