Refactor build pipeline (#1645)
This is a major PR reafctoring the teamcity pipelines.
The goal is:
- to avoid code duplication
- make it easy to extend (add or remove other build steps)
- make it more aligned with other projects (imod-python, coupler, etc)

The structure has changed slighlty
Both the windows and linux projects have a Main buildstep. This
buildstep is a composite buildstep. That means that is isn't building
anything directly, instead it triggers dependend buildsteps. This is the
starting point of the pipeline and only this builstep is directly
triggered from github

It then follows the chain and start the beginning.

Duplicate code have been extracted to Templates. The templates have been
made in such a way that they can both be used by the linux build aswell
as by the windows builds.