Use pixi to fix julia package permissions on Windows (#1710)
Fixes #1603

Given the recent changes in Teamcity, the original project is lost-ish
https://dpcbuild.deltares.nl/buildConfiguration/Ribasim_Ribasim_Windows_FixJuliaArtifactPermissions?branch=continuous_control&buildTypeTab=overview&mode=builds?
As in automatic (?), greyed-out, no build steps or builds visible.
Furthermore, the new Template(s) used is/are platform-agnostic, so
harder to integrate a Windows specific fix.

So the easier way is just to integrate it into pixi, and always running
it at CI time, also useful when it irrevocably happens on Windows
laptops. However, there seems to be no way to have platform specific
tasks (@Hofer-Julian is that correct and maybe worth an issue? I know
you proposed the deno shell 😉), so I now just append an `exit`, ignoring
any exit code of the previous command.

---------

Co-authored-by: Martijn Visser <mgvisser@gmail.com>