Add "add icon" step to windows pipeline (#1712)
Fixes #585 

It is tested on local machine. After ribasim binary is built, if you run
`pixi run add-ribasim-icon`, the ribasim.exe will have an ribasim icon.
The image file is checked in utils/ folder. This "add icon" step is
added to the windows build pipeline, it is right after the build step