Regression tests for ODE solver and integration tests for production model is in place (#1695)
Fixes #462
Fixes #1412 

This PR has a working regression test for ODE solver, and a integration
test for HWS model.

For the regression tests, only 4 models was chosen for regression
testing. In `trivial` model, because it is a very simple model, all the
solvers are tested under 4 different settings of autodiff and sparsity.
No arrow file was used for benchmarking the `trivial` model because it
has only one basin, and the results from different settings/solvers
doesn't vary much. For other 3 models, only the default solver was
tested. It is tested under 4 settings of autodiff and sparsity. The
benchmark arrow files will be in MiniO bucket instead of repository.

For integration test, the benchmark arrow file for HWS model will be in
MiniO bucket as well.

Now if you enter `pixi run test-ribasim-regression` in terminal,
regression tests are run. And `pixi run model-integration-test`,
integration test is run.