Move from ReTestItems to TestItemRunner (#1758)
They are both based on a testitem macro that are quite similar, so it
isn't a big change. The main reason for this is to make sure we can
properly use the VS Code support as documented
[here](https://www.julia-vscode.org/docs/stable/userguide/testitems/).
Right now that only partially works since the two packages have similar
origins.

One downside of this is that we no longer run testitems in parallel by
default. Locally this means going from ~5 to ~8 minutes to run all the
tests, though we should probably verify the timing difference for CI,
and run with `@run_package_tests verbose = true` to see which tests are
taking up most time.

This is prompted by [this Discourse
post](https://discourse.julialang.org/t/new-stable-features-for-the-test-item-framework/118512),
and see also the answer to [my
question](https://discourse.julialang.org/t/new-stable-features-for-the-test-item-framework/118512/9)
about parallel execution.

We should also look into making use of `@testsnippet` and `@testmodule`.