Update reduction factors of resistance nodes (#1796)
These changes are needed for running the `De Dommel` model for Ribasim
NL and were requested by Neeltje. The main change is that
`ManningResistance` has low storage factors now, which is important to
have if the connected basins have different bottom levels.

I made an unified approach for low storage factors of `LinearResistance`
and `ManningResistance`, where the former already had low storage
factors for both connected basins. That however leads to weird behavior
where water cannot flow into an empty basin via the resistance node, so
now I made it so that the reduction factor is based on the storage of
the upstream basin as defined by the flow direction. This feels a bit
discontinuous but I think it's fine.

I also fixed an upstream issue in `FindFirstFunctions`, for now the
manifest points at the branch with the fix:
https://github.com/SciML/FindFirstFunctions.jl/pull/26.