Fix `min_crest_level` migration (#1801)
The migration function from #1788 was wrong because I didn't know that
Ribasim Python would already add the new optional column by the time it
got to the migration. This now drops that one first, to avoid ending up
with `min_upstream_level` twice in one dataframe.

This also makes the other functions a bit simpler by allowing errors
when dropping columns that are no longer needed.