Use pyarrow dtype_backend (#1781)
Fixes #1721
Fixes #1394

The `dtype_backend` part of this is not breaking. The only part that is
technically breaking is that we specify a unit of milliseconds for the
Arrow time type. Previously we used the default nanosecond precision,
which was then truncated to milliseconds in the core. I think it is
better to disallow precision higher than milliseconds if we cannot
distinguish them in the core.