Remove support for resetting cumulatives from BMI (#1852)
The coupler is switching to an approach where they don't need to reset
the cumulatives to 0. Since this complicated the Ribasim code, let's get
rid of it.

The reason for the coupler to move away from resetting cumulatives is
that this possibly broke with #1819.

The dedicated BMI arrays like `vertical_flux_bmi` and
`user_demand.realized_bmi` are removed. The `BMI.get_value_ptr` now just
points to the cumulative state, or if it is not a state (drainage,
precipitation) to `basin.cumulative_drainage` or
`basin.cumulative_precipitation`.

This is breaking because I also renamed these BMI variables to be more
consistent with the current code, where we use the term cumulative
internally.

```
user_demand.realized -> user_demand.inflow
basin.drainage_integrated -> basin.cumulative_drainage
basin.infiltration_integrated ->  basin.cumulative_infiltration
```