Set pixi-version to v0.30.0 (#1870)
Instead of latest. All PRs made since v0.31 came out seem to have pixi
environment related issues. Checking if this will fix it.

#1867 green with v0.30
#1868 red with v0.31
#1869 red with v0.31

This is the error, that somewhat randomly appears during `pixi run`

```
  × /home/runner/work/Ribasim/Ribasim/.pixi/envs/dev/lib/gcc/x86_64-conda-
  │ linux-gnu/14.1.0/libhwasan.so: No such file or directory (os error 2)
```

The failing jobs share that they all run Julia code and are on Ubuntu,
although there are also Julia and Ubuntu jobs that pass.
`gcc_impl_linux-64` is a dependency of rust.