Decrease relative tolerance from 1e-5 to 1e-7 (#1874)
With #1819 the state goes further away from 0 with time. That means we
rely more on the relative tolerance than the absolute tolerance, see the
[SciML
docs](https://docs.sciml.ai/DiffEqDocs/stable/basics/faq/#What-does-tolerance-mean-and-how-much-error-should-I-expect)
on the topic. Hence I found that reducing the `abstol` had no effect,
but reducing the default `reltol` by two orders of magnitude is enough
to get only negligible leaky Terminals with occasional very brief leaks
of 1e-8 m3/s. It is also enough to minimize the infiltration error to
acceptable levels as can be seen in
https://github.com/Deltares/Ribasim/issues/1863#issuecomment-2397302808.

~~For HWS this slows down simulation from 9.1s to 12.3s, which isn't too
bad for a two orders of magnitude reduction in relative tolerance.~~
Rechecking HWS performance gives a speedup! 12.4s for 1e-8, 8.1s for
1e-7, 10.2s for 1e-6. So 1e-7 seems optimal. Not sure what went wrong
last time, but this seems pretty consistent. This is with both absolute
and relative set to the same number.

Fixes #1851
Fixes #1863