Don't rely on juliaup override for version selection
We have

```
install-julia = "juliaup add 1.10.5 && juliaup override unset && juliaup override set 1.10.5"
```

But in #1883 we can see (@evetion can see) that the Julia tests are still run using Julia 1.11, which is the new stable release that came out today.

Rather than trying to find out what exactly goes wrong with the override, this just sets it when we start julia, so we can release.