Use `hatch build` rather than `python -m build`
I got an error that the build module was not found. I tried `pixi add build` but it didn't work and I saw the last release was 3 years ago. Since we already use hatchling it makes sense to use https://hatch.pypa.io/1.12/build/ as well. I already tried it to publish v2024.11.0 and it seems to work fine.