Fix duplicate boundaries in Delwaq generation. (#1903)
Fixes #1901 
Fixes #1902

This disables any dispersion in the Delwaq model, fixing some tracer
intrusions.

#1902 was actually two issues:
- In the graph simplification, we didn't anticipate boundary to boundary
connections (UserDemand to Terminal, in use at IJmuiden). We now just
remove the Terminal node/edge from it.
- UserDemands are the only boundaries that occur twice (when not
connecting back to the same Basin, those are merged to one edge), and
that results in a duplicate boundary name, as it was based on type +
node_id.