Calculate tracer concentrations internally (#1849)
Fixes #1105

This uses the intermediate flows from `update_cumulative_flows` to
update the mass balance in all basins using straightforward
addition/subtraction for each solver timestep. It is a shortcut, assuming piecewise constant concentrations over a timestep, that seems to work fine, but is not a good fit for very large timesteps.