Avoid `curl_easy_setopt` error logs (#1641)
Due to an issue between the curl and openssl libraries the `pixi run
build` tasks failed.
In https://github.com/JuliaLang/Downloads.jl/issues/244 the cause of the
issue is gfound to be environment variables being set that conflict.
This commit adds an environment variable to the tasks failing, clearing
the SSL variables

---------

Co-authored-by: Martijn Visser <mgvisser@gmail.com>