Speed up initialization (#1977)
For lhm_2024_12_3 this speeds up initialization
(`Ribasim.Model(toml_path)`), from 234 seconds to 4 seconds. Large
models were especially slow to initialize since in `create_graph` we
were constructing NodeID structs for each node with a separate call to
the database.

This changes `get_node_ids` to return a `Vector{NodeID}` rather than a
`Vector{Int32}`. This means that the type and index of each ID is now
directly included, leading to simpler code.