Use interpolations to update forcing in Basin (#2006)
Fixes another bullet point from #601, namely:

- use interpolation objects instead of `update_basin`, which still
relies on time being sorted first

Now we can also use `parse_static_and_time` for Basin, bringing our node
types more in line with each other and more code reuse. For this I had
to extend it a bit to support other interpolation types, and support IDs
not being in time nor static. We still add time of new data as tstops to
ensure the solver knows about net data as it happens.

Labeling this as breaking since before this we allowed a node ID to be
in both the static and time tables. In `parse_static_and_time` (the
other node types) we specifically check for this with a clear error
message. I had to update `basic_transient`, to remove the static table
there.