Upload binaries to Minio (#2008)
This is a copy of the Github release, now always triggered, that uploads
to the `latest` folder on our own s3. This is how we did it previously
(year or so back). I've tried to re-use some static credentials from the
AWS connection, so I don't need to introduce new credentials.