Fix minio upload of binaries (#2014)
AWS Shared credentials don't seem to work, as it once to generate a
session token, which Minio can't provide for us (the error on TC is
Minio specific). So I fall back to ENV variables which should be picked
up automatically by the aws command.

Also fixes the name and paths.

Sadly, this is only testable once merged to main AFAIK. I've locally
tested the aws command, but with my own credentials.