Prepare release v2025.1.0 (#2013)
The first release of 2025 makes the Ribasim core more robust and stable,
and speeds up
initialization for large models. The biggest new feature is the
experimental support for
calculating tracers in Ribasim directly, which will hopefully make it
easier to track where
the water resources are heading.

Note that I also pushed a few separate commits with fixes I found during
the release process.