Update to Julia 1.11 (#2031)
We had this earlier with #1914 but that still missed a bugfix that made
package installation more fragile with our antivirus. This switches to
the latest release 1.11.3 that has those fixes.

Since we don't officially support running from Julia this directly
requires 1.11 as the minimum version. I also directly use the new
entrypoint marking.

Somewhat unrelated, but I ran ExplicitImports again and found that we
don't use a bunch of names, so I removed those.

This also sets compat to avoid pulling in ComponentArrays v0.15.23, see
#2028.